task 5: rename importProductFile -> importProductsFile
