+++ import-service/src/functions/index.ts
@@ -1,2 +1,2 @@
-export { default as importProductFile } from "./importProductFile";
+export { default as importProductsFile } from "./importProductsFile";
 export { default as importFileParser } from "./importFileParser";
